Catapult art and bug fixes
The art for the catapult has been added and the code updated to accomodate the changes. Found a bug where Unity would change the player speed based off the performance of the PC. Key inputs will now be handled differently and the code handling physics has been placed back into FixedUpdate(). Also changed some of the values of the player and the dragged objects without altering the feel, in order to fix some issues with dragging.