updated level and added assetes, bug fixes
Finished the parts of the level I can. added in the round runestones, added in the rune stone slot art and fixed the collision bugs that's been plaguing me since the start of the project. No longer will the dragging break unnecisairily and now The level can be built easier.